Real-Time Military Object Detection with YOLOv9
This project utilizes the YOLOv9 model to accurately detect and classify military objects, such as tanks, drones, and warships, in complex visual environments. The model is trained using a large dataset of military imagery to achieve high precision and real-time object recognition. It demonstrates the power of deep learning in enhancing security and surveillance capabilities in military applications.